+++ NXNavigationExtensionDemo/Feature/Common/Models/TableViewSection.m
@@ -42,7 +42,7 @@
                               [TableViewSectionItem itemWithTitle:@"修改导航栏背景颜色" itemType:TableViewSectionItemTypeNavigationBarBackgroundColor],
                               [TableViewSectionItem itemWithTitle:@"修改导航栏背景图片" itemType:TableViewSectionItemTypeNavigationBarBackgroundImage],
                               [TableViewSectionItem itemWithTitle:@"设置导航栏透明" itemType:TableViewSectionItemTypeNavigationBarTransparent],
-                              [TableViewSectionItem itemWithTitle:@"设置导航栏半透明" itemType:TableViewSectionItemTypeNavigationBarTranslucent],
+                              [TableViewSectionItem itemWithTitle:@"实现系统导航栏模糊效果" itemType:TableViewSectionItemTypeLikeSystemBlurNavigationBar],
                               [TableViewSectionItem itemWithTitle:@"设置导航栏底部线条颜色" itemType:TableViewSectionItemTypeNavigationBarShadowColor],
                               [TableViewSectionItem itemWithTitle:@"设置导航栏底部线条图片" itemType:TableViewSectionItemTypeNavigationBarShadowImage],
                               [TableViewSectionItem itemWithTitle:@"自定义返回按钮图片" itemType:TableViewSectionItemTypeNavigationBarCustomBackButtonImage],